refactor utility module boundaries
