+++ include/internal/jit_provider.h
@@ -7,6 +7,5 @@ void peak_jit_provider_enable(void);
 void peak_jit_provider_disable(void);
 gboolean peak_jit_provider_drain_pending(void);
 gboolean peak_jit_provider_drain_pending_force_not_exec_timeout(void);
-gboolean peak_jit_provider_is_enabled(void);
 
 #endif /* __PEAK_JIT_PROVIDER_H */
