+++ src/utils/CMakeLists.txt
@@ -1,12 +1,16 @@
-# Add all source files
-FILE(GLOB sources_utils_c
+# Keep the runtime utility boundary explicit so new or stale files cannot enter
+# the PEAK library merely by appearing in this directory.
+set(sources_utils_c
     env_parser.c
     mpi_utils.c
-    utils.c
+    parent_registry.c
+    process_filter.c
     source_target.c
+    target_config.c
+    timing.c
 )
 
-FILE(GLOB sources_utils_cxx
+set(sources_utils_cxx
     cxx_utils.cpp
 )
