+++ src/utils/parent_registry_internal.h
@@ -1,5 +1,5 @@
-#ifndef PEAK_UTILS_INTERNAL_H
-#define PEAK_UTILS_INTERNAL_H
+#ifndef PEAK_PARENT_REGISTRY_INTERNAL_H
+#define PEAK_PARENT_REGISTRY_INTERNAL_H
 
 #ifdef PEAK_UTILS_TESTING
 typedef enum {
