+++ test/utils/CMakeLists.txt
@@ -13,7 +13,7 @@ set_tests_properties(test_command_filter
 if(CMAKE_SYSTEM_NAME MATCHES "Linux")
     add_executable(test_parent_registry
         test_parent_registry.c
-        ${PROJECT_SOURCE_DIR}/src/utils/utils.c)
+        ${PROJECT_SOURCE_DIR}/src/utils/parent_registry.c)
     target_compile_definitions(test_parent_registry PRIVATE PEAK_UTILS_TESTING=1)
     target_include_directories(test_parent_registry PRIVATE
         ${PROJECT_SOURCE_DIR}/include
